% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bh_plot.R
\name{bh_plot}
\alias{bh_plot}
\title{Create Beverton-Holt curves}
\usage{
bh_plot(this.nc, fg.table, bio.prm, prm.modify, runs.modify, run.dir)
}
\arguments{
\item{run.dir}{}
}
\value{
pdf output curves
}
\description{
Create Beverton-Holt curves
}
