% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_diets_init.R
\name{check_diets_init}
\alias{check_diets_init}
\title{Explore diet interactions}
\usage{
check_diets_init(
  prm.modify,
  run.dir,
  runs.modify,
  fungrouplist,
  atlantis.bgm,
  prm.name,
  this.nc
)
}
\arguments{
\item{this.nc}{}
}
\description{
Explore diet interactions
}
