% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_biomass.R
\name{compare_biomass}
\alias{compare_biomass}
\title{Compare biomass across runs}
\usage{
compare_biomass(fungrouplist, prm.modify, run.dir, maxtimestep)
}
\arguments{
\item{fungrouplist}{list of functional groups}

\item{prm.modify}{list of runs to compare}

\item{run.dir}{directory where runs are stored}

\item{maxtimestep}{maximum desired timestep}
}
\description{
Compare biomass across runs
}
