% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diet_from_init.R
\name{diet_from_init}
\alias{diet_from_init}
\title{Plot diet for predators}
\usage{
diet_from_init(predator, run.dir, this.run, pprey_mat)
}
\arguments{
\item{predator}{}
}
\description{
Plot diet for predators
}
