% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_values.R
\name{extract_values}
\alias{extract_values}
\title{extract species names and param value}
\usage{
extract_values(data)
}
\arguments{
\item{data}{}
}
\value{
ex.values extracted values
}
\description{
extract species names and param value
}
