% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nc_data.R
\name{get_nc_data}
\alias{get_nc_data}
\title{Get nc data}
\usage{
get_nc_data(eachgroup, thisncfile, fungrouplist, runtime, maxtimestep)
}
\arguments{
\item{eachgroup}{each vertebrate group}

\item{thisncfile}{nc file}

\item{fungrouplist}{functional group file}

\item{runtime}{run length}

\item{maxtimestep}{max time step}
}
\value{
thissp.data
}
\description{
Get nc data
}
