% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_output_data.R
\name{get_output_data}
\alias{get_output_data}
\title{Get output data from runs}
\usage{
get_output_data(
  prm.modify,
  runs.modify,
  run.dir,
  run.time,
  fungrouplist,
  this.output.nc,
  maxtimestep
)
}
\arguments{
\item{maxtimestep}{}
}
\description{
Get output data from runs
}
