% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-group_guilds.R
\docType{data}
\name{group_guilds}
\alias{group_guilds}
\title{DATASET_TITLE}
\format{
A data frame with 79 rows and 6 variables:
\describe{
\item{\code{code}}{character COLUMN_DESCRIPTION}
\item{\code{index}}{integer COLUMN_DESCRIPTION}
\item{\code{name}}{character COLUMN_DESCRIPTION}
\item{\code{long_name}}{character COLUMN_DESCRIPTION}
\item{\code{group_type}}{character COLUMN_DESCRIPTION}
\item{\code{guild}}{character COLUMN_DESCRIPTION}
}
}
\usage{
group_guilds
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
