% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_atlantis_runs.R
\name{manage_atlantis_runs}
\alias{manage_atlantis_runs}
\title{Manage and run Atlantis runs}
\usage{
manage_atlantis_runs(sh.file, runs.modify, prm.modify, run.dir)
}
\arguments{
\item{sh.file}{Atlantis run}

\item{runs.modify}{Entries from Google sheet that will be run}

\item{prm.modify}{Google sheet}

\item{run.dir}{Directory where runs are stored}
}
\description{
Manage and run Atlantis runs
}
