% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_prm.R
\name{modify_prm}
\alias{modify_prm}
\title{Modify biology prm}
\usage{
modify_prm(
  eachrun,
  sheet.id,
  sheet.name,
  run.dir,
  prm.modify,
  prm.name,
  atlantis.run
)
}
\arguments{
\item{eachrun}{run number from Google sheet}

\item{sheet.id}{Google sheet id}

\item{sheet.name}{Google sheet name}

\item{run.dir}{Run storage}

\item{prm.modify}{Google sheet with parameter changes}

\item{prm.name}{Name of prm file}

\item{atlantis.run}{Name of run file}
}
\description{
Modify biology prm
}
