% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_biomass.R
\name{plot_biomass}
\alias{plot_biomass}
\title{Plot biomass comparison}
\usage{
plot_biomass(biom.output.file, run.colors, run.dir, runs.modify)
}
\arguments{
\item{biom.output.file}{biomass for all runs}

\item{run.colors}{colors for each run}

\item{runs.modify}{runs that were modified and will be plotted}

\item{run.directory}{directory where runs are stored}
}
\description{
Plot biomass comparison
}
