% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_calibration_runs.R
\name{plot_calibration_runs}
\alias{plot_calibration_runs}
\title{Plot calibration runs}
\usage{
plot_calibration_runs(fungrouplist, prm.modify, run.dir)
}
\arguments{
\item{fungrouplist}{functional group list}

\item{prm.modify}{list of runs to plot from Google sheet}

\item{run.dir}{directory where runs are stored}
}
\description{
Plot calibration runs
}
