% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diets.R
\name{plot_diets}
\alias{plot_diets}
\title{Plot diets}
\usage{
plot_diets(fungrouplist, prm.modify, runs.modify, outdietfile, run.dir)
}
\arguments{
\item{fungrouplist}{list of functional groups}

\item{prm.modify}{file of runs to modify}

\item{runs.modify}{list of run numbers to modify}

\item{outdietfile}{diet file name}

\item{run.dir}{run dir}
}
\description{
Plot diets
}
