% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_guild_biomass.R
\name{plot_guild_biomass}
\alias{plot_guild_biomass}
\title{Plot biomass by guild}
\usage{
plot_guild_biomass(fungrouplist, prm.modify, group_guilds)
}
\arguments{
\item{group_guilds}{}
}
\description{
Plot biomass by guild
}
