% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_param.R
\name{plot_param}
\alias{plot_param}
\title{Plot output variables}
\usage{
plot_param(eachnum, these.runs, folder.paths, varlist)
}
\arguments{
\item{eachnum}{run number}

\item{these.runs}{Runs to plot}

\item{folder.paths}{Paths to runs}

\item{varlist}{List of variables to plot}
}
\description{
Plot output variables
}
