% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_biomass.R
\name{read_biomass}
\alias{read_biomass}
\title{Read biomass}
\usage{
read_biomass(
  eachfolder,
  fg.list,
  folder.paths,
  these.runs,
  maxtimestep,
  biom.output.file
)
}
\arguments{
\item{eachfolder}{folder number}

\item{fg.list}{list of functional groups}

\item{folder.paths}{list of folder paths}

\item{these.runs}{list of runs}

\item{maxtimestep}{maximum desired timestep}
}
\description{
Read biomass
}
