% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_numeric.R
\name{replace_numeric}
\alias{replace_numeric}
\title{Function to replace numeric(0) with NA}
\usage{
replace_numeric(x)
}
\arguments{
\item{x}{}
}
\value{
x
}
\description{
Function to replace numeric(0) with NA
}
