% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nc_data.R
\name{get_nc_data}
\alias{get_nc_data}
\title{Get nc data}
\usage{
get_nc_data(
  eachgroup,
  thisncfile,
  fungrouplist,
  prm.modify,
  maxtimestep,
  outputfrequency
)
}
\arguments{
\item{eachgroup}{each vertebrate group}

\item{thisncfile}{nc file}

\item{fungrouplist}{functional group file}

\item{maxtimestep}{max time step}

\item{runtime}{run length}
}
\value{
thissp.data
}
\description{
Get nc data
}
