% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diets.R
\name{plot_Diets}
\alias{plot_Diets}
\title{Plot diets}
\usage{
plot_Diets(
  fungrouplist,
  prm.modify,
  threshold,
  timeselected,
  starttime,
  outdietfile
)
}
\arguments{
\item{startyear}{}
}
\value{
dietplot
}
\description{
Plot diets
}
